[refac] Rename and reorganize controller packages for better structure and clarity
