+++ edukit-api/src/main/java/com/edukit/api/common/security/handler/JwtAccessDeniedHandler.java
@@ -1,0 +1,36 @@
+package com.edukit.api.common.security.handler;
+
+import com.edukit.api.common.EdukitResponse;
+import com.edukit.core.auth.exception.AuthErrorCode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import org.springframework.http.MediaType;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
+import org.springframework.stereotype.Component;
+
+@Component
+public class JwtAccessDeniedHandler implements AccessDeniedHandler {
+
+    private static final ObjectMapper objectMapper = new ObjectMapper();
+
+    @Override
+    public void handle(final HttpServletRequest request, final HttpServletResponse response,
+                       final AccessDeniedException accessDeniedException) throws IOException {
+        handleException(response);
+    }
+
+    private void handleException(final HttpServletResponse response) throws IOException {
+        setResponse(response);
+    }
+
+    private void setResponse(final HttpServletResponse response) throws IOException {
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        response.setCharacterEncoding(StandardCharsets.UTF_8.name());
+        response.getWriter().write(objectMapper.writeValueAsString(
+                EdukitResponse.fail(AuthErrorCode.FORBIDDEN_MEMBER.getCode(), AuthErrorCode.FORBIDDEN_MEMBER.getMessage())));
+    }
+}
