--- edukit-api/src/main/java/com/edukit/api/security/config/SecurityConfig.java
@@ -1,55 +1,0 @@
-package com.edukit.api.security.config;
-
-import com.edukit.api.security.filter.ExceptionHandlerFilter;
-import com.edukit.api.security.filter.JwtAuthenticationFilter;
-import com.edukit.api.security.handler.JwtAccessDeniedHandler;
-import com.edukit.api.security.handler.JwtAuthenticationEntryPoint;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-    private final CorsConfig corsConfig;
-    private final JwtAuthenticationEntryPoint jwtAuthenticationEntryPoint;
-    private final ExceptionHandlerFilter exceptionHandlerFilter;
-    private final JwtAuthenticationFilter jwtAuthenticationFilter;
-    private final JwtAccessDeniedHandler jwtAccessDeniedHandler;
-
-    @Bean
-    public SecurityFilterChain filterChain(final HttpSecurity http) throws Exception {
-        return http
-                .cors(cors -> cors.configurationSource(corsConfig.corsConfigurationSource()))
-                .csrf(AbstractHttpConfigurer::disable)
-                .formLogin(AbstractHttpConfigurer::disable)
-                .httpBasic(AbstractHttpConfigurer::disable)
-                .sessionManagement(sessionManagementConfigurer -> sessionManagementConfigurer.sessionCreationPolicy(
-                        SessionCreationPolicy.STATELESS))
-                .exceptionHandling(exceptionHandlingConfigurer -> exceptionHandlingConfigurer
-                        .authenticationEntryPoint(jwtAuthenticationEntryPoint)
-                        .accessDeniedHandler(jwtAccessDeniedHandler))
-                .authorizeHttpRequests(auth -> auth
-                        .requestMatchers(SecurityWhitelist.AUTH_WHITELIST).permitAll()
-                        .requestMatchers(SecurityWhitelist.BUSINESS_WHITE_LIST).permitAll()
-                        .requestMatchers(SecurityWhitelist.SWAGGER_WHITE_LIST).permitAll()
-                        .requestMatchers(HttpMethod.OPTIONS, "/**").permitAll()
-                        .requestMatchers("/api/v2/student-records/ai-generate/**").permitAll()
-                        .requestMatchers("/api/v1/admin/**").hasRole("ADMIN")
-                        .requestMatchers("/api/v*/student-records/**").hasAnyRole("TEACHER", "ADMIN")
-                        .anyRequest().authenticated()
-                )
-                .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class)
-                .addFilterBefore(exceptionHandlerFilter, JwtAuthenticationFilter.class)
-                .build();
-    }
-}
